+++ Changes.md
@@ -7,6 +7,7 @@ you spot any mistakes.
 ## NEXT
 
 * Add code `POOL_NOEXIST` in PoolCluster error #846
+* Add `acquireTimeout` pool option to specify a timeout for acquiring a connection #821 #854
 * Add `timeout` option to all sequences #855 #863
 * Default `connectTimeout` to 10 seconds
 * Fix domain binding with `conn.connect`